requirements.txt
anthropic>=0.42.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0

# Dev / Test
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
